refactor(simulation): consolidate shared contracts and output ports
