+++ src/main/java/com/renewsim/backend/simulation_service/shared/application/SimulationDetailsResult.java
@@ -1,4 +1,4 @@
-package com.renewsim.backend.simulation_service.application.shared;
+package com.renewsim.backend.simulation_service.shared.application;
 
 import java.util.List;
