--- src/main/java/com/renewsim/backend/simulation_service/web/dto/ClimateDataDTO.java
@@ -1,9 +1,0 @@
-package com.renewsim.backend.simulation_service.web.dto;
-
-import jakarta.validation.constraints.PositiveOrZero;
-
-public record ClimateDataDTO(
-    @PositiveOrZero double irradiance,
-    @PositiveOrZero double wind,
-    @PositiveOrZero double hydrology
-) {}
