--- src/main/java/com/renewsim/backend/simulation_service/web/dto/SimulationLocationRequestDTO.java
@@ -1,10 +1,0 @@
-package com.renewsim.backend.simulation_service.web.dto;
-
-import jakarta.validation.constraints.DecimalMax;
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.NotNull;
-
-public record SimulationLocationRequestDTO(
-    @NotNull @DecimalMin("-90.0") @DecimalMax("90.0") Double lat,
-    @NotNull @DecimalMin("-180.0") @DecimalMax("180.0") Double lon
-) {}
