+++ src/test/java/com/renewsim/backend/simulation_service/detail/web/SimulationDetailControllerTest.java
@@ -1,8 +1,8 @@
 package com.renewsim.backend.simulation_service.detail.web;
 
 import com.renewsim.backend.auth_service.domain.AuthenticatedUser;
-import com.renewsim.backend.simulation_service.detail.application.GetRealSimulationUseCase;
-import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
+import com.renewsim.backend.simulation_service.detail.application.port.in.GetRealSimulationUseCase;
+import com.renewsim.backend.simulation_service.shared.application.SimulationDetailsResult;
 import com.renewsim.backend.simulation_service.shared.web.dto.SimulationDetailsResponseDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
